Move Legend and Pathway to components
