+++ app/components/Image/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useCallback } from 'react';
 import { Typography, Fade, Button } from '@material-ui/core';
 
-import ImageType from './types';
+import ImageType from './types.d';
 
 import './index.scss';
 
@@ -102,3 +102,5 @@ const Image = ({
 };
 
 export default Image;
+
+export { ImageType };
