+++ app/views/ReportView/components/PathwayAnalysis/index.scss
@@ -17,7 +17,7 @@
       display: flex;
       height: 500px;
       max-width: 800px;
-      margin: 40px;
+      margin:  16px 40px;
     }
   }
 }
